implement print for ansistring
